+++ example/index.jsx
@@ -8,57 +8,75 @@ class App extends Component {
   render() {
     return (
       <div>
-        <PaperRipple tag="h1" color="#ffeee5">
+        <h1>
           React Paper Ripple
-        </PaperRipple>
+          <PaperRipple color="#ffeee5" />
+        </h1>
 
         <h2>Buttons</h2>
         <section>
-          <PaperRipple color="#fff" className="btn">
+          <button color="#fff" className="btn">
             Button
-          </PaperRipple>
-          <PaperRipple color="#f55c03" className="btn btn-link btn-red">
+            <PaperRipple color="#fff" />
+          </button>
+          <button className="btn btn-link btn-red">
             Link
-          </PaperRipple>
+            <PaperRipple color="#f55c03" />
+          </button>
         </section>
 
         <h2>Circle Buttons</h2>
         <section>
-          <PaperRipple color="#f55c03" center className="btn btn-circle btn-red">
-            <i className="icon fa fa-bomb"/>
-          </PaperRipple>
-          <PaperRipple color="#03c8f5" center className="btn btn-circle btn-blue">
-            <i className="icon fa fa-diamond"/>
-          </PaperRipple>
-          <PaperRipple color="#f55c03" center className="btn btn-circle btn-red">
-            <i className="icon fa fa-hand-peace-o "/>
-          </PaperRipple>
+          <button className="btn btn-circle btn-red">
+            <i className="icon fa fa-bomb" />
+            <PaperRipple color="#f55c03" center />
+          </button>
+
+          <button className="btn btn-circle btn-blue">
+            <i className="icon fa fa-diamond" />
+            <PaperRipple color="#03c8f5" center />
+          </button>
+
+          <button className="btn btn-circle btn-red">
+            <i className="icon fa fa-hand-peace-o" />
+            <PaperRipple color="#f55c03" center />
+          </button>
         </section>
 
         <h2>Menu</h2>
         <section>
           <div className="link-list">
-            <PaperRipple color="#dde2e2" className="link">Menu Link</PaperRipple>
-            <PaperRipple color="#dde2e2" className="link">Menu Link</PaperRipple>
-            <PaperRipple color="#dde2e2" className="link">Menu Link</PaperRipple>
+            <div className="link">
+              Menu Link
+              <PaperRipple color="#dde2e2" />
+            </div>
+            <div className="link">
+              Menu Link
+              <PaperRipple color="#dde2e2" />
+            </div>
+            <div className="link">
+              Menu Link
+              <PaperRipple color="#dde2e2" />
+            </div>
           </div>
         </section>
 
         <h2>Image</h2>
         <section>
-          <PaperRipple
+          <div
             color="#fff"
             style={{
               position: 'relative',
-              WebkitUserSelect: 'none'
+              WebkitUserSelect: 'none',
             }}
           >
-            <img src="https://unsplash.it/400/200"/>
-          </PaperRipple>
+            <img src="https://unsplash.it/400/200" />
+            <PaperRipple color="#fff" />
+          </div>
         </section>
       </div>
     )
   }
 }
 
-ReactDOM.render(<App/>, document.getElementById('app'))
+ReactDOM.render(<App />, document.getElementById('app'))